Fix unit tests
